Move LedgerCacheFile implementation into cpp file
